Use Swift Testing for TextBox content sync regression
